requirements.txt
# ── LLM (Groq) ───────────────────────────────────────────
groq>=0.9.0

# ── Data validation ───────────────────────────────────────
pydantic>=2.0

# ── Web framework + async ─────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Task queue ────────────────────────────────────────────
celery[redis]>=5.3.0
redis>=5.0.0

# ── HTTP client (Commerce Agent) ─────────────────────────
httpx>=0.27.0
serpapi>=0.1.5

# ── Image processing (basic) ─────────────────────────────
Pillow>=10.0.0

# ── Vector DB (Style Agent) ───────────────────────────────
chromadb>=0.4.0

# ── Spatial rules (Layout Agent) ─────────────────────────
shapely>=2.0.0

# ── Storage ───────────────────────────────────────────────
cloudinary>=1.40.0

# ── Database ─────────────────────────────────────────────
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv>=1.0.0
numpy>=1.26.0
